[security][CVE-2022-27809] Builtins should always take int64_t, not int
Builtins are allowed to take 32 or 64-bit integers
(int/int32_t versus int64_t). It's not clear why we allow this, as
integers from Hack are always 64-bit. A builtin taking an int will
just have the upper 32-bits truncated. Depending on the value, this
might change a negative number to a positive one, defeating various
input validity checks.
Remove support for 32-bit integers. Change all of the builtins which
took or returned an int to use int64_t instead. This required a few
fixups here and there, mainly around format strings. If the builtin
uses 32-bits ints internally, I left them. At least the truncation is
now obvious.

Test Plan:
New unit test. Beforehand this large negative input would
be turned into a large positive number, bypassing the "input cannot be
negative" check. Now that it takes int64_t, it is correctly recognized
as a negative number.

---
Initializes m_arr in ArrayInitBase

ArrayInitBase's protected constructor does not initialize
m_arr (presumably because a derived class is meant to do so). However
if a derived class' constructor throws, we'll run the ArrayInitBase
destructor, which will try to release m_arr. If m_arr hasn't been
initialized yet, it can contain garbage and we'll try to release
random memory. The protected constructor should initialize m_arr to
nullptr.

Test Plan:
New unit test which previously failed with ASAN builds

+++ hphp/runtime/ext/curl/ext_curl.h
@@ -30,15 +30,15 @@ namespace HPHP {
 ///////////////////////////////////////////////////////////////////////////////
 
 Variant HHVM_FUNCTION(curl_init, const Variant& url = null_string);
-Variant HHVM_FUNCTION(curl_version, int uversion = CURLVERSION_NOW);
-bool HHVM_FUNCTION(curl_setopt, const Resource& ch, int option, const Variant& value);
+Variant HHVM_FUNCTION(curl_version, int64_t uversion = CURLVERSION_NOW);
+bool HHVM_FUNCTION(curl_setopt, const Resource& ch, int64_t option, const Variant& value);
 bool HHVM_FUNCTION(curl_setopt_array, const Resource& ch, const Array& options);
 Variant HHVM_FUNCTION(fb_curl_getopt, const Resource& ch, int64_t opt = 0);
 Variant HHVM_FUNCTION(curl_exec, const Resource& ch);
-Variant HHVM_FUNCTION(curl_getinfo, const Resource& ch, int opt = 0);
+Variant HHVM_FUNCTION(curl_getinfo, const Resource& ch, int64_t opt = 0);
 Variant HHVM_FUNCTION(curl_errno, const Resource& ch);
 Variant HHVM_FUNCTION(curl_error, const Resource& ch);
-String HHVM_FUNCTION(curl_strerror, int code);
+String HHVM_FUNCTION(curl_strerror, int64_t code);
 Variant HHVM_FUNCTION(curl_close, const Resource& ch);
 void HHVM_FUNCTION(curl_reset, const Resource& ch);
 Resource HHVM_FUNCTION(curl_multi_init);
@@ -57,14 +57,13 @@ Variant HHVM_FUNCTION(curl_multi_info_read, const Resource& mh,
                                int64_t& msgs_in_queue);
 Variant HHVM_FUNCTION(curl_multi_close, const Resource& mh);
 bool HHVM_FUNCTION(curl_multi_setopt, const Resource& mh,
-                              int option,
+                              int64_t option,
                               const Variant& value);
 Resource HHVM_FUNCTION(curl_share_init);
 void HHVM_FUNCTION(curl_share_close, const Resource& sh);
 bool HHVM_FUNCTION(curl_share_setopt, const Resource& sh,
-                    int option,
+                    int64_t option,
                     const Variant& value);
 
 ///////////////////////////////////////////////////////////////////////////////
 }
-
