+++ hphp/runtime/ext/url/ext_url.cpp
@@ -58,7 +58,7 @@ String HHVM_FUNCTION(base64_encode, const String& data) {
   return StringUtil::Base64Encode(data);
 }
 
-Variant HHVM_FUNCTION(get_headers, const String& url, int format /* = 0 */) {
+Variant HHVM_FUNCTION(get_headers, const String& url, int64_t format /* = 0 */) {
   Variant c = HHVM_FN(curl_init)();
   HHVM_FN(curl_setopt)(c.toResource(), CURLOPT_URL, url);
   HHVM_FN(curl_setopt)(c.toResource(), CURLOPT_RETURNTRANSFER, true);
@@ -204,7 +204,7 @@ const StaticString s_arg_separator_output("arg_separator.output");
 Variant HHVM_FUNCTION(http_build_query, const Variant& formdata,
                            const Variant& numeric_prefix /* = null_string */,
                            const String& arg_separator /* = null_string */,
-                           int enc_type /* = k_PHP_QUERY_RFC1738 */) {
+                           int64_t enc_type /* = k_PHP_QUERY_RFC1738 */) {
   if (!formdata.isArray() && !formdata.is(KindOfObject)) {
     raise_invalid_argument_warning("formdata: (need Array or Object)");
     return false;
